Cambios en E2E

+++ src/main/java/org/springframework/samples/petclinic/web/VaccineController.java
@@ -7,6 +7,7 @@ import java.util.Map;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.samples.petclinic.model.Product;
 import org.springframework.samples.petclinic.model.Sickness;
 import org.springframework.samples.petclinic.model.Vaccine;
 import org.springframework.samples.petclinic.service.SicknessService;
@@ -70,15 +71,17 @@ public class VaccineController {
 	/* */
 	@GetMapping(path = "/owners/*/pets/*/sicknesses/*/vaccines/{vaccineId}/delete")
 	public String deleteVaccine(@PathVariable("vaccineId") final int vaccineId, final ModelMap modelMap) {
-		String view = "welcome";
-
+		String view;
 		Vaccine vaccine = this.vaccineService.findVaccineById(vaccineId);
-
-		this.vaccineService.delete(vaccine);
-		modelMap.addAttribute("message", "Event sucessfully deleted!");
-
-		modelMap.addAttribute("message", "Event not found!");
-		// view = this.showVaccines(sicknessId, modelMap);
+		if (vaccine != null) {
+			view = "welcome";
+			this.vaccineService.delete(vaccine);
+			modelMap.addAttribute("message", "Event sucessfully deleted!");
+		} else {
+			modelMap.addAttribute("message", "Event not found!");
+			// view = this.showVaccines(sicknessId, modelMap);
+			view = "vaccines/vaccineDontExist";
+		}
 
 		return view;
 
@@ -90,7 +93,7 @@ public class VaccineController {
 	public String editVaccine(@PathVariable("vaccineId") int vaccineId, ModelMap modelMap) {
 
 		Vaccine vaccines = this.vaccineService.findVaccineById(vaccineId);
-		//modelMap.addAttribute("sickness", this.sicknessService.findAll());
+		// modelMap.addAttribute("sickness", this.sicknessService.findAll());
 		Sickness sickness = vaccines.getSickness();
 		modelMap.addAttribute("vaccine", vaccines);
 		modelMap.addAttribute("sickness", sickness);
@@ -98,17 +101,16 @@ public class VaccineController {
 	}
 
 	@PostMapping(value = "/owners/*/pets/*/sicknesses/*/vaccines/{vaccineId}/edit")
-	public String editingVaccine(@Valid Vaccine vaccine, BindingResult result,
-			@PathVariable("vaccineId") int vaccineId , @RequestParam("sickness") final int sicknessId,
-			 ModelMap modelMap) {
+	public String editingVaccine(@Valid Vaccine vaccine, BindingResult result, @PathVariable("vaccineId") int vaccineId,
+			@RequestParam("sickness") final int sicknessId, ModelMap modelMap) {
 
-		String view;																												
+		String view;
 		if (result.hasErrors()) {
 			modelMap.addAttribute("message", "Operation failed!");
 			Sickness enfermedad = this.sicknessService.findSicknessesById(sicknessId);
 			vaccine.setSickness(enfermedad);
 			modelMap.addAttribute("sickness", enfermedad);
-			//modelMap.addAttribute("sickness", this.sicknessService.findAll());
+			// modelMap.addAttribute("sickness", this.sicknessService.findAll());
 			modelMap.addAttribute("vaccine", vaccine);
 			view = "vaccines/updateVaccine";
 
@@ -133,10 +135,10 @@ public class VaccineController {
 	}
 
 	@PostMapping(path = "/vets/newVaccine")
-	public String saveVaccine(@Valid final Vaccine vaccine, final BindingResult result, @RequestParam("sickness") final int sicknessId, 
-			final ModelMap modelMap) {
+	public String saveVaccine(@Valid final Vaccine vaccine, final BindingResult result,
+			@RequestParam("sickness") final int sicknessId, final ModelMap modelMap) {
 		String view = "welcome";
-		
+
 		if (result.hasErrors()) {
 			modelMap.addAttribute("vaccine", vaccine);
 			modelMap.addAttribute("sicknesses", this.sicknessService.findAll());
@@ -146,7 +148,7 @@ public class VaccineController {
 			vaccine.setSickness(enfermedad);
 			this.vaccineService.saveVaccine(vaccine);
 			modelMap.addAttribute("message", "Vaccine succesfully saved!");
-			//view = this.showVaccines(sicknessId, modelMap);
+			// view = this.showVaccines(sicknessId, modelMap);
 		}
 		return view;
 	}
